Fix plugin and widget explorer imports after moving.
